server/requirements-dev.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.6.1
sqlalchemy==2.0.25
PyMySQL==1.1.0
alembic==1.13.1
pytest==8.0.0
httpx==0.26.0
pytest-mock==3.12.0
python-dotenv==1.0.1
bump-pydantic==0.8.0
